Moved the SetLiveNotificationsCommand
It has now been moved to the Twitch package.
