+++ src/main/java/com/beanbeanjuice/command/twitch/SetLiveNotificationsRoleCommand.java
@@ -1,4 +1,4 @@
-package com.beanbeanjuice.command.moderation;
+package com.beanbeanjuice.command.twitch;
 
 import com.beanbeanjuice.main.BeanBot;
 import com.beanbeanjuice.utility.command.CommandContext;
