+++ src/main/java/com/beanbeanjuice/main/BeanBot.java
@@ -8,10 +8,7 @@ import com.beanbeanjuice.command.moderation.*;
 import com.beanbeanjuice.command.moderation.mute.MuteCommand;
 import com.beanbeanjuice.command.moderation.mute.UnMuteCommand;
 import com.beanbeanjuice.command.music.*;
-import com.beanbeanjuice.command.twitch.AddTwitchChannelCommand;
-import com.beanbeanjuice.command.twitch.GetTwitchChannelsCommand;
-import com.beanbeanjuice.command.twitch.RemoveTwitchChannelCommand;
-import com.beanbeanjuice.command.twitch.SetLiveChannelCommand;
+import com.beanbeanjuice.command.twitch.*;
 import com.beanbeanjuice.utility.command.CommandManager;
 import com.beanbeanjuice.utility.guild.GuildHandler;
 import com.beanbeanjuice.utility.helper.GeneralHelper;
